added fund operation modals separately
