+++ packages/ui/components/pages/PoolPage/MarketsList/BorrowModal/Balance.tsx
@@ -1,0 +1,31 @@
+import { Text } from '@chakra-ui/react';
+import { utils } from 'ethers';
+
+import { Column, Row } from '@ui/components/shared/Flex';
+import { SimpleTooltip } from '@ui/components/shared/SimpleTooltip';
+import { useTokenBalance } from '@ui/hooks/useTokenBalance';
+import { MarketData } from '@ui/types/TokensDataMap';
+
+export const Balance = ({ asset }: { asset: MarketData }) => {
+  const { data: myBalance } = useTokenBalance(asset.underlyingToken);
+
+  return (
+    <Column mainAxisAlignment="flex-start" crossAxisAlignment="flex-start" width="100%">
+      <Row width="100%" mainAxisAlignment="flex-end" crossAxisAlignment="center">
+        <Text variant="smText" mr={2}>
+          Wallet Balance:
+        </Text>
+        <SimpleTooltip
+          label={`${myBalance ? utils.formatUnits(myBalance, asset.underlyingDecimals) : 0} ${
+            asset.underlyingSymbol
+          }`}
+        >
+          <Text maxWidth="300px" textOverflow={'ellipsis'} whiteSpace="nowrap" overflow="hidden">
+            {myBalance ? utils.formatUnits(myBalance, asset.underlyingDecimals) : 0}{' '}
+            {asset.underlyingSymbol}
+          </Text>
+        </SimpleTooltip>
+      </Row>
+    </Column>
+  );
+};
