+++ packages/ui/components/pages/PoolPage/MarketsList/BorrowModal/MaxBorrowSlider.tsx
